Be explicit about `activerecord` support
In recent commits, support for `activerecord` versions prior to 5.a (any
versions where the major segment is less than 5) was dropped.

However, a minimum required version in the gemspec file was not added.
That means, if you're using rails 4.2 for example and you are specifying
a requirement for `database_cleaner` that allows prereleases, `bundler`
will choose version 2.0.0.beta2 of the `activerecord` adapter because it
thinks its compatible (no requirement is equivalent to "allow every
version").

We can fix this by being explicit about the minimum version supported.