Add support for ruby version 3.1 and psych 4.0
Ruby 3.1 and above ships with Psych 4.0. Psych throws a `Psych::BadAlias:
Unknown alias: default` error when it loads a `database.yml` which uses YAML
inheritance such as the following:

```yaml
default: &default
  adapter: sqlite3

inherited:
  <<: *default
  database: example.sqlite3
```

We implement the the same pattern as `ActiveSupport::ConfigurationFile` to
circumvent this problem: use `YAML.unsafe_load` when it is available.